pool init&destroy; const strings; some fixes

+++ src/main/java/by/epam/tc/web/service/RoomService.java
@@ -1,5 +1,6 @@
 package by.epam.tc.web.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import by.epam.tc.web.entity.room.Room;
@@ -8,8 +9,8 @@ public interface RoomService {
 
 	List<Room> getAllRooms();
 	Room getRoomByNumber(int roomNumber) throws ServiceException;	
-	int getMinCost();
-	int getMaxCost();
+	BigDecimal getMinCost();
+	BigDecimal getMaxCost();
 	int getMinCapacity();
 	int getMaxCapacity();
 }
